Temp

+++ src/main/java/io/github/zoltus/onecore/player/command/commands/admin/Top.java
@@ -34,7 +34,7 @@ public class Top implements ICommand {
         World w = target.getWorld();
         Location loc = target.getLocation();
         //Gets highest block and adds +1 to y
-        loc.setY(w.getHighestBlockYAt(loc) + 1);
+        loc.setY((double) w.getHighestBlockYAt(loc) + 1);
         //Sets fall distance to 0 so player wont take any fall damage
         target.setFallDistance(0);
         //Resets player velocity so player is falling and teleports to water floating, it wont just fall throught because of the velocity
